refactor: JdbcConnector 만들어 중복 로직 제거
