+++ src/test/java/chess/database/JdbcConnectorTest.java
@@ -1,0 +1,13 @@
+package chess.database;
+
+import static org.junit.jupiter.api.Assertions.assertDoesNotThrow;
+
+import org.junit.jupiter.api.Test;
+
+class JdbcConnectorTest {
+
+    @Test
+    void getConnection() {
+        assertDoesNotThrow(() -> new JdbcConnector().getConnection());
+    }
+}
